WebDavSource.cpp: hijack error 404 to 401 when appropriate.
When doing the first propfindProp, if we get a 404, it might mean
that the username was wrong, so the server gave us a not found
error. It's better to let the user know that, because we don't
have a clear heuristic to determin whether this might have been
a true 404 error.

See also: https://bugs.meego.com/show_bug.cgi?id=17862