GVariant: move common code into libsyncevolution
The same code will also be needed by the upcoming oauth2 backend.  The
header file depends on glib, so don't install it as a public header
file of libsyncevolution (i.e. only allow using it internally).

The original signon code used a custom GVariant auto_ptr because it
needed the simple assignment semantic of "transfers ownership". Now
that the GVariantCXX class is in a shared header file, better make it
identical to the other CXX classes (i.e., a shared pointer). Avoid
returning plain pointers because ownership of those cannot be checked
by the compiler.

The signon case can be handled with GVariantStealCXX, which by default
takes ownership in the constructor and the default assignment
operator.