oauth2: refresh token -> oauth2
That the backend is based on a refresh token is an implementation
detail; it might even change at some point if we figure out how
to do OAuth2 internally.

Better use the shorter "oauth2" name in the "username" property.

Also better align .so name and global variable names/defines in
configure and Makefile with the name of the backend.