datatypes: raw text items with minimal conversion
When using "raw/text/calendar" or "raw/text/vcard" as SyncEvolution
"databaseFormat", all parsing and conversion is skipped. The backend's
data is identical to the item data in the engine. Finding duplicates
in a slow sync is very limited when using these types because the entire
item data must match exactly.

This is useful for the file backend when the goal is to store an exact copy
of what a peer has or for limited, read-only backends (PBAP).

The implementation uses text profiles with just a single field which
intentionally is named "data", because that is the key that
SyncSourceSerialize also uses to transfer its
data. SyncSourceSerialize can thus be used unchanged, we only need to
skip the MAKETEXTWITHPROFILE/PARSETEXTWITHPROFILE by not providing a
profile name.

A different item type might be even more efficient, but this already
goes a long way towards eliminating overhead for parsing/generating.

In a refresh-from-remote local sync of 1000 contacts with photo data
using the file backend, this reduces the number of CPU cycles from
2861476184 to 1967552611 on the server side.